[Refactor] exterior color 패키지 구조 변경 및 usecase 단위 port, adapter 추가 (#320)
* refactor: usecase별 port, adaptor 추가

* refactor: 불필요하게 긴 변수명 수정
