+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/color_exterior/controller/ExteriorColorController.java
@@ -12,7 +12,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import softeer.be_my_car_master.api.color_exterior.dto.request.GetExteriorColorsRequest;
 import softeer.be_my_car_master.api.color_exterior.dto.response.GetExteriorColorsResponse;
-import softeer.be_my_car_master.api.color_exterior.usecase.GetExteriorColorsUseCase;
+import softeer.be_my_car_master.api.color_exterior.usecase.get_exterior_colors.GetExteriorColorsUseCase;
 import softeer.be_my_car_master.global.exception.BindingParamException;
 import softeer.be_my_car_master.global.response.Response;
 
@@ -26,15 +26,15 @@ public class ExteriorColorController {
 	@GetMapping("/exterior-colors")
 	@Operation(summary = "트림에서 선택가능한 외장 색상 목록을 반환합니다")
 	public Response<GetExteriorColorsResponse> getExteriorColors(
-		@Valid @ParameterObject GetExteriorColorsRequest getExteriorColorsRequest,
+		@Valid @ParameterObject GetExteriorColorsRequest request,
 		BindingResult bindingResult
 	) {
 		if (bindingResult.hasErrors()) {
 			throw new BindingParamException(bindingResult.getFieldErrors());
 		}
 
-		Long trimId = getExteriorColorsRequest.getTrimId();
-		GetExteriorColorsResponse getExteriorColorsResponse = getExteriorColorsUseCase.execute(trimId);
-		return Response.createSuccessResponse(getExteriorColorsResponse);
+		Long trimId = request.getTrimId();
+		GetExteriorColorsResponse response = getExteriorColorsUseCase.execute(trimId);
+		return Response.createSuccessResponse(response);
 	}
 }
