+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/color_exterior/usecase/get_exterior_colors/GetExteriorColorsUseCase.java
@@ -1,10 +1,9 @@
-package softeer.be_my_car_master.api.color_exterior.usecase;
+package softeer.be_my_car_master.api.color_exterior.usecase.get_exterior_colors;
 
 import java.util.List;
 
 import lombok.RequiredArgsConstructor;
 import softeer.be_my_car_master.api.color_exterior.dto.response.GetExteriorColorsResponse;
-import softeer.be_my_car_master.api.color_exterior.usecase.port.ExteriorColorPort;
 import softeer.be_my_car_master.domain.color_exterior.ExteriorColor;
 import softeer.be_my_car_master.global.annotation.UseCase;
 
@@ -12,10 +11,10 @@ import softeer.be_my_car_master.global.annotation.UseCase;
 @RequiredArgsConstructor
 public class GetExteriorColorsUseCase {
 
-	private final ExteriorColorPort exteriorColorPort;
+	private final GetExteriorColorsPort port;
 
 	public GetExteriorColorsResponse execute(Long trimId) {
-		List<ExteriorColor> selectableExteriorColors = exteriorColorPort.findSelectableExteriorColorsByTrimId(trimId);
+		List<ExteriorColor> selectableExteriorColors = port.findExteriorColorsByTrim(trimId);
 		return GetExteriorColorsResponse.from(selectableExteriorColors);
 	}
 }
